backend/requirements.txt
asgiref==3.11.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
lxml==6.1.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
PyJWT==2.12.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-magic-bin==0.4.14
pytz==2026.1.post1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.2
